BF: Adjust to new behavior, simplify earlier changes
